third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-fetch-blocked.html

<!DOCTYPE html>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script>
if (window.testRunner) {
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", false);
}

var test = async_test("Opens a HTTPS window that loads insecure data via Fetch API. We should trigger a mixed content callback and block the load.");

window.addEventListener("message", test.step_func(function (e) {
    assert_equals(e.data, "DONE");
    test.done();
}), false);
window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-fetch.html");
</script>
